Release kti-5.2.
It would appear that, when running a batch, on the KTI branch simulations finish simulation faster than they can be added to the registry runner queue. This appears to create a sort of 'race like condition' in which the runner registry reports that there are no more simulations left, before all the simulations have been added, leaving both GUI and program in slightly strange state.

Instead of starting the polling timer when any model is added to the runner registry, we instead add all the models to the registry and then start the timer. It's possible this problem could reappear when a batch whilst the registry runner is already dealing with another batch run. however this tail case would require more effort to solve, simply patch the problem to get the thing out the door.